server/requirements.txt
Flask
python-dotenv
PyJWT
flask-cors
